refactor: split out decimal related code from `slice_operation.rs` to `slice_decimal_operation.rs`
