+++ crates/proof-of-sql/src/base/database/mod.rs
@@ -9,6 +9,8 @@ pub use column::{Column, ColumnField, ColumnRef, ColumnType};
 
 mod slice_operation;
 
+mod slice_decimal_operation;
+
 mod column_type_operation;
 pub use column_type_operation::{
     try_add_subtract_column_types, try_divide_column_types, try_multiply_column_types,
